Audio tweaks (#103)
keira: liltracker: fix calculated buffer size, fix int_t sizes for pattern/event indexes
sdk: add LILKA_NO_AUDIO_HELLO & LILKA_NO_BUZZER_HELLO
doc: add info about audio flags

+++ firmware/keira/src/apps/liltracker/liltracker.cpp
@@ -692,7 +692,7 @@ int LilTrackerApp::drawElement(
 }
 
 void LilTrackerApp::startPreview(
-    Track* track, page_t* page, int32_t requestedChannelIndex, int32_t requestedEventIndex
+    Track* track, page_t* page, int8_t requestedChannelIndex, uint16_t requestedEventIndex
 ) {
     for (int channelIndex = 0; channelIndex < CHANNEL_COUNT; channelIndex++) {
         Pattern* pattern = track->getPattern(page->patternIndices[channelIndex]);
