+++ firmware/keira/src/apps/liltracker/liltracker.h
@@ -16,7 +16,7 @@ private:
         const char* text, int16_t x, int16_t y, lilka::Alignment hAlign, lilka::Alignment vAlign, bool editing,
         bool focused, uint16_t color
     );
-    void startPreview(Track* track, page_t* page, int32_t requestedChannelIndex, int32_t requestedEventIndex);
+    void startPreview(Track* track, page_t* page, int8_t requestedChannelIndex, uint16_t requestedEventIndex);
     void alert(String title, String message);
     bool confirm(String title, String message);
     String filePicker(bool isSave);
